🛠️ : move Buy required items into Process component
what: move required-item purchasing UI/logic from process detail page into shared Process component.

why: embedded process cards (quests/items) need the same Buy required items action.

how to test: npm run test:root -- frontend/src/components/__tests__/Process.spec.ts frontend/src/pages/process/[slug]/__tests__/ProcessView.spec.ts && npm run lint
